+++ app/Component/View/mobile.js
@@ -5,6 +5,7 @@
 
 import React, {Component} from 'react';
 import {ScrollView} from 'react-native';
+import omit from 'lodash/omit';
 
 export default class ReactorsViewMobile extends Component {
   props: $reactors$Core$props;
@@ -14,8 +15,15 @@ export default class ReactorsViewMobile extends Component {
   }
 
   render() {
+    const props = omit(this.props.style);
+
     return (
-      <ScrollView ref="__internalView" {...this.props}>
+      /* $FlowFixMe - we don't have react */
+      <ScrollView
+        ref="__internalView"
+        {...props}
+        contentContainerStyle={this.props.style}
+        >
         {this.props.children}
       </ScrollView>
     );